JUnit Ver 4 requires assertEquals method to have explicit types- assertEquals(ContextAnnotator.MIDDLE_SCOPE, (int)ca.scopes.get(0)); -notice the type case in parameter #2
git-svn-id: https://svn.apache.org/repos/asf/incubator/ctakes/trunk/cTAKES@1385536 13f79535-47bb-0310-9956-ffa450edef68

+++ NE contexts/test/src/edu/mayo/bmi/uima/context/ContextAnnotatorTests.java
@@ -512,9 +512,9 @@ public class ContextAnnotatorTests {
 		ContextAnnotator ca = new ContextAnnotator();
 		ca.parseScopeOrder(new String[] { "MIDDLE", "LEFT", "RIGHT", "ALL" });
 		assertEquals(4, ca.scopes.size());
-		assertEquals(ContextAnnotator.MIDDLE_SCOPE, ca.scopes.get(0));
-		assertEquals(ContextAnnotator.LEFT_SCOPE, ca.scopes.get(1));
-		assertEquals(ContextAnnotator.RIGHT_SCOPE, ca.scopes.get(2));
-		assertEquals(ContextAnnotator.ALL_SCOPE, ca.scopes.get(3));
+		assertEquals(ContextAnnotator.MIDDLE_SCOPE, (int)ca.scopes.get(0));
+		assertEquals(ContextAnnotator.LEFT_SCOPE, (int)ca.scopes.get(1));
+		assertEquals(ContextAnnotator.RIGHT_SCOPE, (int)ca.scopes.get(2));
+		assertEquals(ContextAnnotator.ALL_SCOPE, (int)ca.scopes.get(3));
 	}
 }
